Finish moving the projects tests over
