+++ api/tests/unit/projects/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+
+from segments.models import Segment
+
+
+@pytest.fixture()
+def segments(project):
+    segments = []
+    for i in range(3):
+        segments.append(
+            Segment.objects.create(name=f"Test Segment {i}", project=project)
+        )
+    return segments
